refactor final ui design and code structure with optimization
