+++ package.json
@@ -23,7 +23,7 @@
     "react-dom": "^18.3.1",
     "react-json-to-table": "^0.1.7",
     "react-redux": "^9.1.2",
-    "react-scripts": "5.0.1",
+    "react-scripts": "^5.0.0",
     "redux": "^5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
@@ -33,12 +33,6 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
   },
   "browserslist": {
     "production": [
