+++ src/components/CenteredBox.tsx
