+++ src/components/GraphDisplay.tsx
