+++ src/components/MessageBubble.tsx
