+++ src/components/TextDisplay.tsx
