+++ src/hooks/useReduxHooks.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { RootState, AppDispatch } from '../app/Store';
+import type { RootState, AppDispatch } from '../redux/store/Store';
 
 // Custom hooks to use Redux dispatch and selector with proper types
 export const useAppDispatch = () => useDispatch<AppDispatch>();
