+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import store from './app/Store';
+import store from './redux/store/Store';
 import App from './App';
 
 ReactDOM.render(
