+++ src/redux/store/Store.ts
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import chatReducer from '../features/chatSlice'; 
+import chatReducer from '../slices/chatSlice'; 
 
 const store = configureStore({
   reducer: {
